requirements copy.txt
absl-py==2.1.0
asgiref==3.7.2
asn1crypto==1.5.1
astunparse==1.6.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli==1.1.0
cachetools==5.3.2
celery==5.4.0
certifi==2023.7.22
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==41.0.5
cssselect2==0.7.0
daphne==4.1.2
deepface==0.0.92
defusedxml==0.8.0rc2
Django==5.0
django-cors-headers==4.3.0
django-sslserver==0.22
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
et-xmlfile==1.1.0
fido2==1.1.3
filelock==3.15.4
fire==0.6.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.46.0
freetype-py==2.3.0
gast==0.6.0
gdown==5.2.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
gunicorn==22.0.0
html5lib==1.1
hyperlink==21.0.0
idna==3.4
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
keras==3.4.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.0.8
mtcnn==0.1.1
namex==0.0.8
numpy==1.26.2
oauthlib==3.2.2
opencv-contrib-python-headless==4.8.1.78
opencv-python==4.10.0.84
opencv-python-headless==4.8.1.78
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.12.1
oscrypto==1.3.0
packaging==24.1
protobuf==4.25.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydyf==0.8.0
Pygments==2.18.0
pyHanko==0.21.0
pyhanko-certvalidator==0.26.2
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyphen==0.14.0
pypng==0.20220715.0
psycopg2-binary==2.9.9
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.8
reportlab==4.2.2
requests==2.31.0
requests-oauthlib==1.3.1
retina-face==0.0.17
rich==13.7.1
rlPyCairo==0.3.0
rsa==4.9
service-identity==24.1.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
soupsieve==2.5
sqlparse==0.4.4
svglib==1.5.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
tf_keras==2.17.0
tinycss2==1.2.1
tqdm==4.66.5
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.8.0
tzlocal==5.2
uritools==4.0.2
urllib3==2.1.0
webencodings==0.5.1
Werkzeug==3.0.3
wrapt==1.16.0